requirements.txt
pandas
matplotlib
rdkit-pypi
scipy
numpy